backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
requests==2.32.3
gunicorn==22.0.0

